refactor(graph_runtime_state): Move runtime protocol to runtime package
