+++ api/core/workflow/graph/__init__.py
@@ -1,6 +1,7 @@
+from core.workflow.runtime import ReadOnlyGraphRuntimeState, ReadOnlyVariablePool
+
 from .edge import Edge
 from .graph import Graph, NodeFactory
-from .graph_runtime_state_protocol import ReadOnlyGraphRuntimeState, ReadOnlyVariablePool
 from .graph_template import GraphTemplate
 from .read_only_state_wrapper import ReadOnlyGraphRuntimeStateWrapper, ReadOnlyVariablePoolWrapper
