+++ api/core/workflow/graph_engine/response_coordinator/coordinator.py
@@ -18,7 +18,7 @@ from core.workflow.enums import NodeExecutionType, NodeState
 from core.workflow.graph import Graph
 from core.workflow.graph_events import NodeRunStreamChunkEvent, NodeRunSucceededEvent
 from core.workflow.nodes.base.template import TextSegment, VariableSegment
-from core.workflow.runtime.variable_pool import VariablePool
+from core.workflow.runtime import VariablePool
 
 from .path import Path
 from .session import ResponseSession
