+++ api/core/workflow/runtime/__init__.py
@@ -1,4 +1,11 @@
 from .graph_runtime_state import GraphRuntimeState
+from .graph_runtime_state_protocol import ReadOnlyGraphRuntimeState, ReadOnlyVariablePool
 from .variable_pool import VariablePool, VariableValue
 
-__all__ = ["GraphRuntimeState", "VariablePool", "VariableValue"]
+__all__ = [
+    "GraphRuntimeState",
+    "ReadOnlyGraphRuntimeState",
+    "ReadOnlyVariablePool",
+    "VariablePool",
+    "VariableValue",
+]
