+++ api/tests/unit_tests/core/workflow/entities/test_graph_runtime_state.py
@@ -2,8 +2,7 @@ from time import time
 
 import pytest
 
-from core.workflow.runtime.graph_runtime_state import GraphRuntimeState
-from core.workflow.runtime.variable_pool import VariablePool
+from core.workflow.runtime import GraphRuntimeState, VariablePool
 
 
 class TestGraphRuntimeState:
