+++ api/tests/unit_tests/core/workflow/graph_engine/test_mock_nodes_template_code.py
@@ -17,8 +17,7 @@ class TestMockTemplateTransformNode:
     def test_mock_template_transform_node_default_output(self):
         """Test that MockTemplateTransformNode processes templates with Jinja2."""
         from core.workflow.entities import GraphInitParams
-        from core.workflow.runtime import GraphRuntimeState
-        from core.workflow.runtime.variable_pool import VariablePool
+        from core.workflow.runtime import GraphRuntimeState, VariablePool
 
         # Create test parameters
         graph_init_params = GraphInitParams(
@@ -78,8 +77,7 @@ class TestMockTemplateTransformNode:
     def test_mock_template_transform_node_custom_output(self):
         """Test that MockTemplateTransformNode returns custom configured output."""
         from core.workflow.entities import GraphInitParams
-        from core.workflow.runtime import GraphRuntimeState
-        from core.workflow.runtime.variable_pool import VariablePool
+        from core.workflow.runtime import GraphRuntimeState, VariablePool
 
         # Create test parameters
         graph_init_params = GraphInitParams(
@@ -140,8 +138,7 @@ class TestMockTemplateTransformNode:
     def test_mock_template_transform_node_error_simulation(self):
         """Test that MockTemplateTransformNode can simulate errors."""
         from core.workflow.entities import GraphInitParams
-        from core.workflow.runtime import GraphRuntimeState
-        from core.workflow.runtime.variable_pool import VariablePool
+        from core.workflow.runtime import GraphRuntimeState, VariablePool
 
         # Create test parameters
         graph_init_params = GraphInitParams(
@@ -200,8 +197,7 @@ class TestMockTemplateTransformNode:
         """Test that MockTemplateTransformNode processes templates with variables."""
         from core.variables import StringVariable
         from core.workflow.entities import GraphInitParams
-        from core.workflow.runtime import GraphRuntimeState
-        from core.workflow.runtime.variable_pool import VariablePool
+        from core.workflow.runtime import GraphRuntimeState, VariablePool
 
         # Create test parameters
         graph_init_params = GraphInitParams(
@@ -267,8 +263,7 @@ class TestMockCodeNode:
     def test_mock_code_node_default_output(self):
         """Test that MockCodeNode returns default output."""
         from core.workflow.entities import GraphInitParams
-        from core.workflow.runtime import GraphRuntimeState
-        from core.workflow.runtime.variable_pool import VariablePool
+        from core.workflow.runtime import GraphRuntimeState, VariablePool
 
         # Create test parameters
         graph_init_params = GraphInitParams(
@@ -329,8 +324,7 @@ class TestMockCodeNode:
     def test_mock_code_node_with_output_schema(self):
         """Test that MockCodeNode generates outputs based on schema."""
         from core.workflow.entities import GraphInitParams
-        from core.workflow.runtime import GraphRuntimeState
-        from core.workflow.runtime.variable_pool import VariablePool
+        from core.workflow.runtime import GraphRuntimeState, VariablePool
 
         # Create test parameters
         graph_init_params = GraphInitParams(
@@ -399,8 +393,7 @@ class TestMockCodeNode:
     def test_mock_code_node_custom_output(self):
         """Test that MockCodeNode returns custom configured output."""
         from core.workflow.entities import GraphInitParams
-        from core.workflow.runtime import GraphRuntimeState
-        from core.workflow.runtime.variable_pool import VariablePool
+        from core.workflow.runtime import GraphRuntimeState, VariablePool
 
         # Create test parameters
         graph_init_params = GraphInitParams(
@@ -471,8 +464,7 @@ class TestMockNodeFactory:
     def test_code_and_template_nodes_mocked_by_default(self):
         """Test that CODE and TEMPLATE_TRANSFORM nodes are mocked by default (they require SSRF proxy)."""
         from core.workflow.entities import GraphInitParams
-        from core.workflow.runtime import GraphRuntimeState
-        from core.workflow.runtime.variable_pool import VariablePool
+        from core.workflow.runtime import GraphRuntimeState, VariablePool
 
         # Create test parameters
         graph_init_params = GraphInitParams(
@@ -513,8 +505,7 @@ class TestMockNodeFactory:
     def test_factory_creates_mock_template_transform_node(self):
         """Test that MockNodeFactory creates MockTemplateTransformNode for template-transform type."""
         from core.workflow.entities import GraphInitParams
-        from core.workflow.runtime import GraphRuntimeState
-        from core.workflow.runtime.variable_pool import VariablePool
+        from core.workflow.runtime import GraphRuntimeState, VariablePool
 
         # Create test parameters
         graph_init_params = GraphInitParams(
@@ -565,8 +556,7 @@ class TestMockNodeFactory:
     def test_factory_creates_mock_code_node(self):
         """Test that MockNodeFactory creates MockCodeNode for code type."""
         from core.workflow.entities import GraphInitParams
-        from core.workflow.runtime import GraphRuntimeState
-        from core.workflow.runtime.variable_pool import VariablePool
+        from core.workflow.runtime import GraphRuntimeState, VariablePool
 
         # Create test parameters
         graph_init_params = GraphInitParams(
